compute: walk an expensive index peek off the serving worker
Moves an expensive fast-path index peek's walk off the timely worker that
received it, so a long scan no longer delays the peeks queued behind it. This is
the layer the ones beneath it were built for.

A peek runs its first slice inline under a small budget, sized so point lookups
finish there and nothing else does. If it completes, the peek never leaves the
worker. If it outruns the budget it is offloaded to a task that walks the rest
on the blocking pool, holding a permit from a process-wide semaphore. The walk
stays on its blocking thread, checking for cancellation every few thousand
positions, and returns to the runtime only to answer: a round trip per slice
costs two thread wakes, which is more than the slice itself on a small machine.
Cost is measured rather than predicted, so a skewed point lookup over a hot key
needs no special case: it enters as a point lookup, overruns, and offloads.

The offload is for a scan that suspended with its budget spent, never for one
holding a full batch. An offloaded task has nowhere to write a batch until the
stash becomes a state transition of the same scan, and stepping a batch-ready
scan returns without spending fuel or advancing, so offloading one would spin. A
scan that fills a batch while offloaded hands back and the worker starts the
existing stash walk.

The permit is owned by the running task rather than by the pending peek, so
releasing it coincides with releasing the batches it accounts for. Permits
default to this runtime's worker count, which preserves the ceiling that exists
today: a peek blocking its worker already costs one core per worker. Excess scans
queue in the semaphore, so a peek storm costs queue entries rather than threads.

Cancellation needs no new mechanism at any stage. Removing the pending peek drops
the result channel's receiver, and the walk observes a closed sender at its next
cancellation check; a scan still waiting for a permit leaves the queue the same
way.
Permits release on drop, including on panic.

The peeks awaiting a turn on the worker and the peeks a driver has taken over sit
in separate queues, because they need opposite treatment. The former all draw on
one per-activation aggregate that does not refill within an activation, so the
sweep serves them from the front and stops at the first peek the budget cannot
serve. The latter draw no budget, since their work is not on the worker, so the
sweep polls all of them. A peek that gets no turn keeps its place and is served
first on the next activation.

Off by default in production and on in the test configuration. With the switch
off nothing is offloaded and placement is unchanged.

Both substrate counters are pre-resolved, so
`mz_index_peek_walks_total{substrate="offloaded"}` reads zero rather than being
absent, which keeps "the offload changed nothing" distinguishable from "the
offload never engaged". Reading `mz_index_peek_total_seconds` across the flag is
misleading: an offloaded peek contributes only its inline slice, which the inline
budget bounds, so the expensive peeks leave a bounded sample where they used to
leave the whole walk. `mz_index_peek_offload_seconds` and the per-phase
histograms are the honest pairing.

+++ src/compute/src/compute_state/peek_scan/tests.rs
@@ -42,7 +42,7 @@ type OksHandle = PaddedTrace<RowRowAgent<Timestamp, Diff>>;
 const RESUMPTION_BOUND: usize = 100;
 
 /// The rows `value` values yield, in the order the ok walk produces them.
-fn rows(values: impl IntoIterator<Item = u8>) -> Vec<Row> {
+fn rows(values: impl IntoIterator<Item = u64>) -> Vec<Row> {
     values.into_iter().map(row).collect()
 }
 
@@ -57,7 +57,7 @@ fn row_size() -> usize {
 }
 
 /// The rows and counts a completed scan over `values` hands back.
-fn expected(values: impl IntoIterator<Item = u8>) -> RowBatch {
+fn expected(values: impl IntoIterator<Item = u64>) -> RowBatch {
     values
         .into_iter()
         .map(|value| (row(value), NonZeroI64::new(1).expect("non-zero")))
